project: Merged FFmpeg Encoders Step 2

+++ source/obs/gs/gs-sampler.cpp
@@ -19,6 +19,7 @@
 
 #include "gs-sampler.hpp"
 #include <stdexcept>
+#include "utility.hpp"
 
 gs::sampler::sampler()
 {
@@ -97,7 +98,8 @@ void gs::sampler::set_border_color(uint32_t v)
 void gs::sampler::set_border_color(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
 {
 	_dirty                     = true;
-	_sampler_info.border_color = a << 24 | r << 16 | g << 8 | b;
+	_sampler_info.border_color = (static_cast<uint32_t>(a) << 24) | (static_cast<uint32_t>(r) << 16)
+								 | (static_cast<uint32_t>(g) << 8) | static_cast<uint32_t>(b);
 }
 
 uint32_t gs::sampler::get_border_color()
